Feat: 결제 방법 수정하기 페이지에 적용하는 토글 함수 구현
