+++ client/front/lib/fixerrand/fixerrand.dart
@@ -4,6 +4,7 @@ import 'dart:developer';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
+import 'package:front/fixerrand/fixiscash.dart';
 import 'package:front/fixerrand/fixminimap.dart';
 import 'package:geolocator/geolocator.dart';
 
@@ -34,6 +35,7 @@ class _FixErrandState extends State<FixErrand> {
   late String detail;
   late int reward;
   late String status;
+  late bool isCash;
 
   final int maxTitleLength = 20; // 제목 최대 길이 설정
 
@@ -47,11 +49,6 @@ class _FixErrandState extends State<FixErrand> {
   bool isDetailAddressEnabled = true;
   bool isPriceEnabled = true;
   bool isRequestEnabled = true;
-
-  // 결제 방법 토글 버튼 변수 선언
-  bool isAccountTransfer = true;
-  late bool isCash = false;
-  late List<bool> isSelected2 = [isAccountTransfer, isCash];
 
   @override
   void initState() {
@@ -74,6 +71,7 @@ class _FixErrandState extends State<FixErrand> {
     longitude = widget.errands['longitude'];
     due = widget.errands['due'];
     createdDate = widget.errands['createdDate'];
+    isCash = widget.errands['isCash'];
   }
 
     @override
@@ -111,19 +109,6 @@ class _FixErrandState extends State<FixErrand> {
       // 비밀번호 확인 입력란의 텍스트 변경 감지하여 확인 버튼의 활성화 상태 업데이트
       setState(() {
         isRequestEnabled = requestController.text.isNotEmpty;
-      });
-    }
-
-    // 토글 버튼(결제 방법)
-    void toggleSelect2(int newindex) {
-      setState(() {
-        for (int index = 0; index < isSelected2.length; index++) {
-          if (index == newindex) {
-            isSelected2[index] = true;
-          } else {
-            isSelected2[index] = false;
-          }
-        }
       });
     }
 
@@ -476,58 +461,8 @@ class _FixErrandState extends State<FixErrand> {
                           ),
                         ),
                       ),
-                      Container(
-                        width: 119,
-                        height: 31,
-                        // 토글 버튼 만들기
-                        margin: EdgeInsets.only(right: 86),
-                        child: ToggleButtons(
-                          color: Color(0xff2E2E2E),
-                          // 선택되지 않은 버튼 텍스트 색상
-
-                          borderColor: Colors.grey,
-                          // 토글 버튼 테두리 색상
-                          borderWidth: 0.5,
-                          borderRadius: BorderRadius.circular(5.0),
-
-                          selectedColor: Color(0xffC77749),
-                          // 선택된 버튼 텍스트 색상
-                          fillColor: Color(0xffFFFFFF),
-                          // 선택된 버튼 배경색
-                          selectedBorderColor: Color(0xffC77749),
-                          // 선택된 버튼 테두리 색상
-
-                          // renderBorder: false,
-                          children: <Widget>[
-                            Padding(
-                              padding: EdgeInsets.symmetric(horizontal: 11),
-                              child: Text(
-                                '계좌이체',
-                                style: TextStyle(
-                                  fontFamily: 'Pretendard',
-                                  fontWeight: FontWeight.w400,
-                                  fontSize: 13,
-                                  letterSpacing: 0.01,
-                                ),
-                              ),
-                            ),
-                            Padding(
-                              padding: EdgeInsets.symmetric(horizontal: 14),
-                              child: Text(
-                                '현금',
-                                style: TextStyle(
-                                  fontFamily: 'Pretendard',
-                                  fontWeight: FontWeight.w400,
-                                  fontSize: 13,
-                                  letterSpacing: 0.01,
-                                ),
-                              ),
-                            ),
-                          ],
-                          isSelected: isSelected2,
-                          onPressed: toggleSelect2,
-                        ),
-                      ),
+                      // 결제 방법
+                      FixIsCash(isCash: isCash,),
                     ],
                   ),
                 ),
@@ -612,7 +547,7 @@ class _FixErrandState extends State<FixErrand> {
                     ),
                   ),
                 ),
-                // 작성 완료 버튼 만들기
+                // 수정 완료 버튼 만들기
                 Container(
                   margin: EdgeInsets.only(left: 20.0, right: 20.0, top: 18),
                   child: ElevatedButton(
